Move preferences views into directory
