+++ bookwyrm/views/__init__.py
@@ -1,4 +1,5 @@
 """ make sure all our nice views are available """
+# site admin
 from .admin.announcements import Announcements, Announcement, delete_announcement
 from .admin.dashboard import Dashboard
 from .admin.federation import Federation, FederatedServer
@@ -19,13 +20,17 @@ from .admin.reports import (
 )
 from .admin.site import Site
 from .admin.user_admin import UserAdmin, UserAdminList
+# user preferences
+from .preferences.change_password import ChangePassword
+from .preferences.edit_user import EditUser
+from .preferences.delete_user import DeleteUser
+# misc views
 from .author import Author, EditAuthor
 from .block import Block, unblock
 from .books import Book, EditBook, ConfirmEditBook
 from .books import upload_cover, add_description, resolve_book
 from .directory import Directory
 from .discover import Discover
-from .edit_user import EditUser, DeleteUser
 from .editions import Editions, switch_edition
 from .feed import DirectMessage, Feed, Replies, Status
 from .follow import follow, unfollow
@@ -47,7 +52,7 @@ from .reading import delete_readthrough, delete_progressupdate
 from .reading import ReadingStatus
 from .register import Register, ConfirmEmail, ConfirmEmailCode, resend_link
 from .rss_feed import RssFeed
-from .password import PasswordResetRequest, PasswordReset, ChangePassword
+from .password import PasswordResetRequest, PasswordReset
 from .search import Search
 from .shelf import Shelf
 from .shelf import create_shelf, delete_shelf
