+++ bookwyrm/views/preferences/edit_user.py
@@ -1,9 +1,8 @@
-""" edit or delete ones own account"""
+""" edit your own account """
 from io import BytesIO
 from uuid import uuid4
 from PIL import Image
 
-from django.contrib.auth import logout
 from django.contrib.auth.decorators import login_required
 from django.core.files.base import ContentFile
 from django.shortcuts import redirect
@@ -11,7 +10,7 @@ from django.template.response import TemplateResponse
 from django.utils.decorators import method_decorator
 from django.views import View
 
-from bookwyrm import forms, models
+from bookwyrm import forms
 
 
 # pylint: disable=no-self-use
@@ -37,35 +36,6 @@ class EditUser(View):
         save_user_form(form)
 
         return redirect("user-feed", request.user.localname)
-
-
-# pylint: disable=no-self-use
-@method_decorator(login_required, name="dispatch")
-class DeleteUser(View):
-    """delete user view"""
-
-    def get(self, request):
-        """delete page for a user"""
-        data = {
-            "form": forms.DeleteUserForm(),
-            "user": request.user,
-        }
-        return TemplateResponse(request, "preferences/delete_user.html", data)
-
-    def post(self, request):
-        """les get fancy with images"""
-        form = forms.DeleteUserForm(request.POST, instance=request.user)
-        # idk why but I couldn't get check_password to work on request.user
-        user = models.User.objects.get(id=request.user.id)
-        if form.is_valid() and user.check_password(form.cleaned_data["password"]):
-            user.deactivation_reason = "self_deletion"
-            user.delete()
-            logout(request)
-            return redirect("/")
-
-        form.errors["password"] = ["Invalid password"]
-        data = {"form": form, "user": request.user}
-        return TemplateResponse(request, "preferences/delete_user.html", data)
 
 
 def save_user_form(form):
